requirements.txt
torch>=2.1
transformers==4.50.0
accelerate
Pillow
xformers
